+++ DepRegAttributes.ExampleLibrary/RegisteredClasses/KeyedServices/KeyedWithConst.cs
@@ -1,0 +1,9 @@
+﻿using DepRegAttributes.ExampleLibrary.RegisteredClasses.TaggedServices;
+
+namespace DepRegAttributes.ExampleLibrary.RegisteredClasses.KeyedServices;
+
+[RegisterTransient(Key = Const.Value)]
+public class KeyedWithConst
+{
+
+}