+++ DepRegAttributes.ExampleLibrary/RegisteredClasses/KeyedServices/KeyedWithInterface.cs
@@ -1,0 +1,10 @@
+﻿namespace DepRegAttributes.ExampleLibrary.RegisteredClasses.KeyedServices;
+
+public interface IKeyedWithInterface
+{
+}
+
+[RegisterTransient<IKeyedWithInterface>(Key = "Value")]
+public class KeyedWithInterface : IKeyedWithInterface
+{
+}