+++ DepRegAttributes.ExampleLibrary/RegisteredClasses/KeyedServices/KeyedWithInternalPrivateConst.cs
@@ -1,0 +1,7 @@
+﻿namespace DepRegAttributes.ExampleLibrary.RegisteredClasses.KeyedServices;
+
+[RegisterTransient(Key = Value)]
+public class KeyedWithInternalPrivateConst
+{
+    private const string Value = nameof(Value);
+}